material route fix

+++ route/material.js
@@ -21,9 +21,9 @@ router.post("/list", verifyAdmin, (req, res, next) => {
     d.name, d.description, d.fileId,
     f.name as filename, f.size, f.mimetype, f.path
     FROM materials m
-    INNER JOIN blocks b ON b.id = m.blockId
-    INNER JOIN documents d ON d.id = m.documentId
-    INNER JOIN files f ON f.id = d.fileId
+    LEFT JOIN blocks b ON b.id = m.blockId
+    LEFT JOIN documents d ON d.id = m.documentId
+    LEFT JOIN files f ON f.id = d.fileId
     WHERE m.blockId = ?`;
   const params = [blockId];
   db.all(query, params, (err, rows) => {
@@ -60,11 +60,11 @@ router.post("/list", verifyAdmin, (req, res, next) => {
 router.get("/get/:id", verifyAdmin, (req, res, next) => {
   const query = `SELECT m.id, m.blockId, m.documentId, 
     d.name, d.description, d.fileId,
-    f.name, f.size, f.mimetype, f.path
+    f.name as filename, f.size, f.mimetype, f.path
     FROM materials m
-    INNER JOIN blocks b ON b.id = m.blockId
-    INNER JOIN documents d ON d.id = m.documentId
-    INNER JOIN files f ON f.id = d.fileId
+    LEFT JOIN blocks b ON b.id = m.blockId
+    LEFT JOIN documents d ON d.id = m.documentId
+    LEFT JOIN files f ON f.id = d.fileId
     WHERE m.id = ?`;
   const params = [req.params.id]
   db.get(query, params, (err, row) => {
@@ -77,7 +77,22 @@ router.get("/get/:id", verifyAdmin, (req, res, next) => {
     }
     res.json({
       success: true,
-      data: row
+      data: {
+        id: row.id,
+        blockId: row.blockId,
+        document: {
+          id: row.documentId,
+          name: row.name,
+          description: row.description,
+          file: {
+            id: row.fileId,
+            name: row.filename,
+            size: row.size,
+            mimetype: row.mimetype,
+            path: row.path
+          }
+        }
+      }
     })
   });
 });
@@ -143,16 +158,64 @@ router.post("/create", verifyAdmin, (req, res, next) => {
   });
 });
 router.patch("/update/:id", verifyAdmin, (req, res, next) => {
-  const query = `SELECT m.id, m.blockId, m.documentId, 
-    d.name, d.description, d.fileId,
-    f.name, f.size, f.mimetype, f.path
-    FROM materials m
-    INNER JOIN blocks b ON b.id = m.blockId
-    INNER JOIN documents d ON d.id = m.documentId
-    INNER JOIN files f ON f.id = d.fileId
-    WHERE m.id = ?`;
-  const params = [req.params.id]
-  db.get(query, params, (err, material) => {
+  const errors = []
+  if (!req.body.blockId) {
+    errors.push("No blockId specified");
+  }
+  if (!req.body.document) {
+    errors.push("No document specified");
+  }
+  if (!req.body.document.name) {
+    errors.push("No name specified");
+  }
+  if (!req.body.document.description) {
+    errors.push("No description specified");
+  }
+  if (!req.body.document.file.id) {
+    errors.push("No fileId specified");
+  }
+  if (errors.length) {
+    res.status(400).json({
+      success: false,
+      message: errors.join(", ")
+    })
+    return;
+  }
+
+  const data = {
+    id: req.params.id,
+    blockId: req.body.blockId,
+    document: req.body.document
+  };
+  const updateDocuments = () => {
+    const query = `UPDATE documents SET
+        name = COALESCE(?,name), 
+        description = COALESCE(?,description), 
+        fileId = COALESCE(?,fileId)
+        WHERE id = ?`;
+    const params = [data.document.name, data.document.description, data.document.file.id, data.document.id];
+    db.run(query, params, function (err, result) {
+      if (err){
+        res.status(400).json({
+          success: false,
+          message: err.message
+        })
+        return;
+      }
+      res.json({
+        success: true,
+        changes: this.changes,
+        data: data
+      })
+    });
+  }
+
+  const query = `SELECT f.id, f.path
+    FROM documents d
+    LEFT JOIN files f ON f.id = d.fileId
+    WHERE d.id = ?`;
+  const params = [data.document.id];
+  db.get(query, params, (err, file) => {
     if (err) {
       res.status(400).json({
         success: false,
@@ -160,10 +223,12 @@ router.patch("/update/:id", verifyAdmin, (req, res, next) => {
       })
       return;
     }
-    if (material.file.id) {
-      fs.unlinkSync(`./${material.document.file.path}`);
+    if (file.id !== data.document.file.id) {
+      if (fs.existsSync(`./${file.path}`)) {
+        fs.unlinkSync(`./${file.path}`);
+      }
       const query = `DELETE FROM files WHERE id = ?`;
-      const params = [material.document.file.id];
+      const params = [file.id];
       db.run(query, params, function (err, result) {
         if (err){
           res.status(400).json({
@@ -172,64 +237,43 @@ router.patch("/update/:id", verifyAdmin, (req, res, next) => {
           })
           return;
         }
-        const query = `UPDATE documents SET
-          name = COALESCE(?,name), 
-          description = COALESCE(?,description), 
-          fileId = COALESCE(?,fileId)
-          WHERE id = ?`;
-        const params = [material.document.name, material.document.description, material.document.file.id, material.id];
-        db.run(query, params, function (err, result) {
-          if (err){
-            res.status(400).json({
-              success: false,
-              message: err.message
-            })
-            return;
-          }
-          res.json({
-            success: true,
-            changes: this.changes,
-            data: material
-          })
-        });
+        updateDocuments();
       });
     } else {
-      const query = `UPDATE documents SET
-        name = COALESCE(?,name), 
-        description = COALESCE(?,description), 
-        fileId = COALESCE(?,fileId)
-        WHERE id = ?`;
-      const params = [material.document.name, material.document.description, material.document.file.id, material.id];
-      db.run(query, params, function (err, result) {
-        if (err){
-          res.status(400).json({
-            success: false,
-            message: err.message
-          })
-          return;
-        }
-        res.json({
-          success: true,
-          changes: this.changes,
-          data: material
-        })
-      });
+      updateDocuments();
     }
   });
 })
 router.delete("/delete", verifyAdmin, (req, res, next) => {
   const ids = req.body.ids;
+  let changes = 0;
   ids?.forEach((id) => {
     const query = `SELECT m.id, m.blockId, m.documentId, 
-      d.name, d.description, d.fileId,
-      f.name, f.size, f.mimetype, f.path
-      FROM materials m
-      INNER JOIN blocks b ON b.id = m.blockId
-      INNER JOIN documents d ON d.id = m.documentId
-      INNER JOIN files f ON f.id = d.fileId
-      WHERE m.id = ?`;
+    d.name, d.description, d.fileId,
+    f.name as filename, f.size, f.mimetype, f.path
+    FROM materials m
+    LEFT JOIN blocks b ON b.id = m.blockId
+    LEFT JOIN documents d ON d.id = m.documentId
+    LEFT JOIN files f ON f.id = d.fileId
+    WHERE m.id = ?`;
     const params = [id]
-    db.get(query, params, (err, material) => {
+    db.get(query, params, (err, row) => {
+      const material = {
+        id: row.id,
+        blockId: row.blockId,
+        document: {
+          id: row.documentId,
+          name: row.name,
+          description: row.description,
+          file: {
+            id: row.fileId,
+            name: row.filename,
+            size: row.size,
+            mimetype: row.mimetype,
+            path: row.path
+          }
+        }
+      };
       if (err) {
         res.status(400).json({
           success: false,
@@ -237,11 +281,13 @@ router.delete("/delete", verifyAdmin, (req, res, next) => {
         })
         return;
       }
-      fs.unlinkSync(`./${material.document.file.path}`);
+      if (fs.existsSync(`./${material.document.file.path}`)) {
+        fs.unlinkSync(`./${material.document.file.path}`);
+      }
       const query = `DELETE FROM files WHERE id = ?`;
       const params = [material.document.file.id];
       db.run(query, params, function (err, result) {
-        if (err){
+        if (err) {
           res.status(400).json({
             success: false,
             message: err.message
@@ -251,7 +297,7 @@ router.delete("/delete", verifyAdmin, (req, res, next) => {
         const query = `DELETE FROM documents WHERE id = ?`;
         const params = [material.document.id];
         db.run(query, params, function (err, result) {
-          if (err){
+          if (err) {
             res.status(400).json({
               success: false,
               message: err.message
@@ -261,22 +307,23 @@ router.delete("/delete", verifyAdmin, (req, res, next) => {
           const query = `DELETE FROM materials WHERE id = ?`;
           const params = [material.id];
           db.run(query, params, function (err, result) {
-            if (err){
+            if (err) {
               res.status(400).json({
                 success: false,
                 message: err.message
               })
               return;
             }
-            res.json({
-              success: true,
-              changes: this.changes,
-            })
+            changes += this.changes;
           });
         });
       });
     });
   });
+  res.json({
+    success: true,
+    changes: changes,
+  })
 })
 
 module.exports = router
